+++ frontend/src/App.js
@@ -1,9 +1,9 @@
 import React from 'react';
 import {observer} from 'mobx-react';
 import UserStore from './stores/UserStore';
-import LoginForm from './LoginForm';
-import InputField from './InputField';
-import SubmitButton from './SubmitButton';
+import LoginForm from './components/LoginForm';
+import InputField from './components/InputField';
+import SubmitButton from './components/SubmitButton';
 import './App.css';
 
 class App extends React.Component {
