+++ frontend/src/components/LoginForm.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import InputField from './InputField';
 import SubmitButton from './SubmitButton';
-import UserStore from'./stores/UserStore';
+import UserStore from'../stores/UserStore';
 
 class LoginForm extends React.Component {
